+++ internal/provider/resource_sdwan_attach_feature_device_template.go
@@ -20,6 +20,7 @@ package provider
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/CiscoDevNet/terraform-provider-sdwan/internal/provider/helpers"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -141,7 +142,10 @@ func (r *AttachFeatureDeviceTemplateResource) Read(ctx context.Context, req reso
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", state.Id.String()))
 
 	err := state.readVariables(ctx, r.client)
-	if err != nil {
+	if err != nil && strings.Contains(err.Error(), "StatusCode 400") {
+		resp.State.RemoveResource(ctx)
+		return
+	} else if err != nil {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s", err))
 		return
 	}