+++ docs/guides/changelog.md
@@ -32,6 +32,7 @@ description: |-
 - Add `sdwan_cellular_controller_feature_template` resource and data source
 - BREAKING CHANGE: Increases the minimum version for all profile parcel resource and data sources to 20.12.0
 - Fix issue when changing multiple already deployed templates including device templates
+- Fix error when refreshing state of `sdwan_attach_feature_device_template` resource and device template does no longer exist
 
 ## 0.3.6